requirements.txt
# 测试框架
pytest>=7.0.0
pytest-html>=3.1.0
pytest-xdist>=2.5.0

# HTTP 请求
requests>=2.28.0

# 配置解析
PyYAML>=6.0

# 数据处理
pandas>=1.5.0
jsonschema>=4.0.0

# 日志
colorlog>=6.7.0

# SQL 解析
sqlparse>=0.4.0

# 测试报告（可选）
allure-pytest>=2.12.0

# 数据库连接（可选）
mysql-connector-python>=8.0.0

# 打包工具（打包 exe 时需要）
pyinstaller>=5.0.0